Fix flow inference empty token seq

+++ cosyvoice/flow/flow.py
@@ -251,7 +251,7 @@ class CausalMaskedDiffWithXvec(torch.nn.Module):
         # concat text and prompt_text
         token, token_len = torch.concat([prompt_token, token], dim=1), prompt_token_len + token_len
         mask = (~make_pad_mask(token_len)).unsqueeze(-1).to(embedding)
-        token = self.input_embedding(torch.clamp(token, min=0)) * mask
+        token = self.input_embedding(torch.clamp(token, min=0).long()) * mask
 
         # text encode
         if finalize is True:
@@ -300,7 +300,7 @@ class CausalMaskedDiffWithXvec(torch.nn.Module):
         embedding = embedding.expand(batch_size, -1)
         token, token_len = torch.concat([prompt_token.expand(batch_size, -1), token], dim=1), prompt_token_len + token_len
         mask = (~make_pad_mask(token_len)).unsqueeze(-1).to(embedding)
-        token = self.input_embedding(torch.clamp(token, min=0)) * mask
+        token = self.input_embedding(torch.clamp(token, min=0).long()) * mask
 
         # text encode
         if finalize is True:
